Separate app.py into handlers
