+++ savoten/handler/__init__.py
@@ -1,0 +1,2 @@
+from .event_api import EventAPI  # noqa F401
+from .event_view import event_view  # noqa F401
